fix(storage): apply the document rules at the constructor, not only from_dict
Review flagged two constructor paths reproducing defects already fixed at the
document boundary — `diagnostics="parse error"` split into eleven
one-character entries, and `attributes=("ab",)` unpacked as the valid-looking
pair `("a", "b")`, colliding with an occurrence that really held it. Rather
than fix the two named, I audited every guard for a constructor twin and found
six unguarded paths: provenance and diagnostics on `FactAvailability`,
attribute rows and the container/producer site fields on `OccurrenceId`, and
the family/entity keys on `declare()`/`override()`.

The generalizable rule, now stated where the values enter: validating at the
document boundary is not enough for a publicly constructible type. Every rule
this module enforces on a parsed document is enforced on a direct call too.

Two things the gates caught in my own work, both worth recording:

The first attempt added a *second* `__post_init__` to `FactAvailability`,
which silently shadowed the existing one and dropped its `status`/`confidence`
checks. That guard's own pre-existing test failed immediately — adding
validation is exactly how you shadow validation. The rules are merged into one
method, and a test now pins the status guard alongside the new ones.

`test_identity.py` then crossed the 1200-line test cap, so it is split the
same way `test_availability.py` was: `test_identity_documents.py` takes the
boundary contracts — what a document or a constructor may contain, and whether
a record survives a round trip — and the original keeps retention, keys,
ordering, and who judges a contradiction.

Regression tests verified failing against the pre-fix tree (13 of 21).
